Added the 03 Workshop - very simple multi-page application that was aimed to improve my abilities to initiate express app, design middleware, work with Handlebars templates.
I designed search to work the following way: if there is user input in the field - the app is taking the value into consideration and checks whether there are matches only for search criteria that are FILLED. For example: If "from" is 5 and "to" and "search" empty - only the models with difficulty greater than 5 will be displayed.

If "search" = "gan" and "from"=1 => models that have "gan" in their name or description and with difficulty greater than 5 will be displayed.

Pending - to make checks for correct user data and proper error handling 

Also the way I filter the models that match the criteria seems a bit clumsy - will think of better solution